Model Checking Before Simulation
RegularSimulator.py_simulate now asks the CSimInterface object to call check_interface(). This function should be filled in by different interfaces to check that they are internally ready to simulate. In the case of the CSimModelInterface, this has been done by checking to make sure the model is initialized. If this function is not filled in, there will be a warning but no error.

Additional changes include minor cleanups like removing some old unused variables and warnings / prints. Ready to be pushed onto master.