SBML Annotations and refactoring SBML Code
Bioscrape can now read annotations in SBML kinetic_laws to tell it which kind of propensity to use. This allows for SBML models to be compiled into more efficient bioscrape models.

In order to make this work, legacy SBML support was removed and all SBML related functionalities were moved to a new folder: sbmlutil.py.

Example notebooks were changed to work with the new code.